+++ vscode-lox/CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change function snippet detail from "function declaration" to "function".
 - Change class snippet detail from "class declaration" to "class".
 - Add placeholders to all snippets where appropriate.
+- Add function name to detail of completion and outline items
 
 ### Removed